Collapse update_edition to updater.edition
